requirements.txt
Pillow==10.2.0
langchain==0.3.25
openai==1.78.0
python-dotenv==1.0.1
langgraph==0.0.26
langchain-core==0.3.59